requirements.txt
matplotlib>=3.7.0
numpy>=1.24.0
pillow>=10.0.0
